lab1/app/requirements.txt
antiberty==0.1.1
biopython==1.81
fastapi==0.110.0
joblib==1.3.2
numpy==1.26.4
pandas==2.1.4
pydantic==2.6.4
scikit-learn==1.3.2
torch==2.2.2
uvicorn[standard]==0.23.2
